Extract ImageTypeIcon to a new component
